Completed the 2020 day problem. Changed solution output to be Vec[u64] instead of Vec[u32] since the day 9 problem has larger numbers. Made the Parseable implementation for u32 be generic for unsigned numbers to be able to handle u64 as well.

+++ src/aoc_2020/day_06.rs
@@ -90,7 +90,7 @@ pub const SOLUTION: Solution = Solution {
         ];
 
         // Processing
-        let answers = part_questions.iter().map(|v| v.iter().map(|q| q.len() as u32).sum()).collect();
+        let answers = part_questions.iter().map(|v| v.iter().map(|q| q.len() as u64).sum()).collect();
         
         Ok(answers)
     }
